Organize into an src folder
