+++ index.html
@@ -5,7 +5,7 @@
   <meta charset='utf-8'>
   <meta name='viewport' content='width=device-width, initial-scale=1'>
   <link rel='shortcut icon' href='https://mapbox.com/favicon.ico' type='image/x-icon'>
-  <link href='main.css' rel='stylesheet'>
+  <link href='./src/styles/main.css' rel='stylesheet'>
 </head>
 <body>
   <div id='app'></div>
