+++ index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app.js';
+import App from './src/components/app.js';
 
 ReactDOM.render((<App />), document.getElementById('app'));
