+++ package.json
@@ -15,7 +15,8 @@
     "build": "browserify index.js | uglifyjs > bundle.js",
     "test": "jest",
     "lint": "eslint .",
-    "lint-css": "stylelint main.css"
+    "fix": "eslint --fix .",
+    "lint-css": "stylelint src/styles/main.css"
   },
   "engines": {
     "yarn": "*"
