+++ test/app.test.js
@@ -1,7 +1,6 @@
 /* eslint-disable no-undef */
-
 import React from 'react';
-import App from '../components/app';
+import App from '../src/components/app';
 import { mount } from 'enzyme';
 
 test('welcomes the user to React', function() {
